refactor(infra): 环境与工具层公共接口统一
仅替换 environment/ 和 utils/ 模块对外部模块的子模块导入为公共接口导入，
同时补充目标包 __init__.py 缺失的 re-export。

- clients/__init__.py: 新增 GitHubClient, BackendProtocol, find_repo_root
- models/__init__.py: 新增 OrchestraConfig
- utils/__init__.py: 新增 AUTOMATED_MARKERS, GENERIC_AGENT_MARKER_PATTERN, STARTING_TIMEOUT_SECONDS
- config/__init__.py: 新增 load_orchestra_config
- services/__init__.py: 新增 get_manager_usernames

- worktree.py: SQLiteClient, OrchestraConfig (lazy)
- worktree_pr_mixin.py: GitHubClient, SQLiteClient, OrchestraConfig (lazy)
- worktree_lifecycle.py: SQLiteClient, OrchestraConfig (lazy)
- session_registry.py: BackendProtocol, SQLiteClient, STARTING_TIMEOUT_SECONDS, find_repo_root (lazy)
- worktree_support.py: OrchestraConfig (lazy)

- comment_utils.py: load_orchestra_config, get_manager_usernames, AUTOMATED_MARKERS, GENERIC_AGENT_MARKER_PATTERN
- codeagent_helpers.py: VibeConfig

- 类型检查: mypy PASS
- 测试: 128 tests PASS
- Lint: ruff PASS
- 集成验证: 所有 re-export 路径导入成功

Closes #1716
