+++ src/vibe3/clients/__init__.py
@@ -1,7 +1,8 @@
 """Vibe3 clients layer."""
 
-from vibe3.clients.git_client import GitClient
+from vibe3.clients.git_client import GitClient, find_repo_root
 from vibe3.clients.github_client import GitHubClient
+from vibe3.clients.protocols import BackendProtocol
 from vibe3.clients.serena_client import (
     SerenaClient,
     count_references,
@@ -10,10 +11,12 @@ from vibe3.clients.serena_client import (
 from vibe3.clients.sqlite_client import SQLiteClient
 
 __all__ = [
+    "BackendProtocol",
     "GitClient",
     "GitHubClient",
     "SerenaClient",
     "SQLiteClient",
     "count_references",
     "extract_function_names",
+    "find_repo_root",
 ]
