+++ src/vibe3/config/__init__.py
@@ -1,6 +1,7 @@
 """Config package."""
 
 from vibe3.config.loader import get_config, load_config, reload_config
+from vibe3.config.orchestra_settings import load_orchestra_config
 from vibe3.config.settings import (
     CodeLimitsConfig,
     CodePathsConfig,
@@ -17,6 +18,7 @@ from vibe3.config.settings import (
 __all__ = [
     "get_config",
     "load_config",
+    "load_orchestra_config",
     "reload_config",
     "VibeConfig",
     "CodeLimitsConfig",
