+++ src/vibe3/environment/session_registry.py
@@ -5,10 +5,9 @@ from typing import Any
 
 from loguru import logger
 
-from vibe3.clients.protocols import BackendProtocol
-from vibe3.clients.sqlite_client import SQLiteClient
+from vibe3.clients import BackendProtocol, SQLiteClient
 from vibe3.environment.session_naming import build_session_name
-from vibe3.utils.constants import STARTING_TIMEOUT_SECONDS
+from vibe3.utils import STARTING_TIMEOUT_SECONDS
 
 # All supported execution roles (L1/L2/L3 chains)
 # Exported from here to avoid circular imports with constants module
@@ -312,7 +311,7 @@ class SessionRegistryService:
             return
 
         try:
-            from vibe3.clients.git_client import find_repo_root
+            from vibe3.clients import find_repo_root
             from vibe3.environment.worktree import WorktreeManager
             from vibe3.environment.worktree_context import WorktreeContext
