+++ src/vibe3/environment/worktree_pr_mixin.py
@@ -11,12 +11,11 @@ from typing import TYPE_CHECKING, Optional
 
 from loguru import logger
 
-from vibe3.clients.github_client import GitHubClient
-from vibe3.clients.sqlite_client import SQLiteClient
+from vibe3.clients import GitHubClient, SQLiteClient
 from vibe3.environment.worktree_context import WorktreeContext
 
 if TYPE_CHECKING:
-    from vibe3.models.orchestra_config import OrchestraConfig
+    from vibe3.models import OrchestraConfig
 
 
 class WorktreePRMixin:
