+++ src/vibe3/utils/__init__.py
@@ -1,3 +1,13 @@
 """Utility modules for Vibe3."""
 
-__all__ = []
+from vibe3.utils.constants import (
+    AUTOMATED_MARKERS,
+    GENERIC_AGENT_MARKER_PATTERN,
+    STARTING_TIMEOUT_SECONDS,
+)
+
+__all__ = [
+    "AUTOMATED_MARKERS",
+    "GENERIC_AGENT_MARKER_PATTERN",
+    "STARTING_TIMEOUT_SECONDS",
+]
